backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv

chromadb
sentence-transformers
Pillow
numpy
